Fixed some bugs associated with the corners heuristic
Bugs ixed:
1. When running the corner heuristic with the tinyCorners maze, there
used to be an "empty step" in the path. This caused an exception to be
thrown. I removed all empty steps from the path. This fixed the problem.

2. The heuristic worked only when the list of corners comes in the same
order in which we have to visit the corners. If the order was different
- an exception happened. I fixed this by finding the closest corner to
pacman's current position, and using the manhattan distance for that
corner.

3. I played with coefficients to lower the number of expanded nodes with
the cornersHeuristic. I lowered them down to 451. I can lower them down
to 361 but this may impact other mazes. I found that the coefficient I
have right now works with all mazes.

4. Sometimes problems illegal commands were generated from the statement
explored[i][0] = [thisSuccessor, successorFn]  in aStarHelper, so I
commented it out.